Fix dirty pixels bug
When drawing a rectangle and fooling with the cursor around the
original clicked point, dirty pixels were left and could appear in
the resulting screenshot + they were left when scrot exited.

I think this was because the XOR rectangle had trouble with a 1 px
height/width. Ensuring a rectangle is only drawn/erased when it's at
least larger than a 5x5 px square seems to fix the bug.